ai-service/requirements.txt
flask
python-dotenv
google-generativeai
Flask-Cors

